Aula4

Description

LI1

Documentation

type Matriz a = [[a]]

Matriz sobre um tipo de dados

trocaPU

Arguments

:: Matriz a
Matriz original não vazia

-> Matriz a
Matriz resultante, com a primeira e a última linha trocada

Esta função recebe uma matriz não vazia, troca a primeira linha com a última

trocaCO

Arguments

:: Matriz a
Matriz original não vazia

-> Matriz a
Matriz resultante, com a primeira coluna trocada com a última coluna

Função principal

Esta função recebe uma matriz não vazia, e troca a primeira coluna com a última

trocaCOAux :: [a] -> [a]

Função auxiliar

elemento_matriz

Arguments

:: (Int, Int)
Posição dada

-> Matriz Int
Matriz dada

-> Int
Elemento da matriz

Esta função dada a posicao devolve o elemento

devolve_linha :: Int -> Matriz Int -> [Int]

devolve_coluna :: Int -> [Int] -> Int