Ficha3

Description

LI1

Documentation

remove

Arguments

:: [String]
Lista de palavras

-> Char
Character a remover

-> [String]
Lista de palavras atualizada

Esta função que recebe uma lista de strings e remove todas as strings iniciadas por um dado caractere

add

Arguments

:: [(Int, Int)]
Lista de pares inteiros

-> Int
Valor dado para adicionar

-> [(Int, Int)]
Lista de pares inteiros resultante

Esta função recebe uma lista de pares de inteiros e adiciona um valor dado à primeira componente de cada par

maior

Arguments

:: [(Int, Int)]
Lista de pares inteiros (obs: lista não pode ser vazia)

-> Int
Maior valor da segunda componente

Esta função recebe uma lista, não vazia, de pares de inteiros e calcula qual o maior valor da segunda componente

type Nome = String

Este tipo de dados representa o nome de uma pessoa

type Coordenada = (Double, Double)

Este tipo de dados representa as coordenadas de uma pessoa

data Movimento

Constructors

N
Norte

S
Sul

E
Este

W
Oeste

type Movimentos = [Movimento]

data PosicaoPessoa

Este tipo de dados representa a posição de uma pessoa

Constructors

Pos

Fields

Nome

Nome da pessoa

Coordenada

Coordenadas da pessoa

posicoesM

Arguments

:: [PosicaoPessoa]
Lista de posições de pessoas

-> Movimento
Movimento a ser efetuado

-> [PosicaoPessoa]
Lista de posições de pessoas atualizadas

Esta função recebe uma lista de posições de pessoas, atualize essa lista depois de todas executarem um movimento dado

posicao :: PosicaoPessoa -> Movimentos -> PosicaoPessoa

Esta função recebe a posição de uma pessoa depois de executar uma sequência de movimentos

posicoesMs

Arguments

:: [PosicaoPessoa]
Lista de posições de pessoas

-> Movimentos
Movimentos a ser efetuados

-> [PosicaoPessoa]
Lista de posições de pessoas atualizadas

Esta função recebe uma lista de posições de pessoas, atualize essa lista depois de todas as pessoas executarem uma mesma sequência de movimentos

pessoaNorte

Arguments

:: [PosicaoPessoa]
Lista de posições de pessoas

-> Maybe PosicaoPessoa
Pessoa mais a norte

Esta função recebe uma lista de posições de pessoas e indica qual a posição mais a norte (em caso de igualdade, escolha uma posição)